Move to dedicated morph target module
