+++ crates/bevy_render/src/mesh/mesh/mod.rs
@@ -1,4 +1,5 @@
 mod conversions;
+mod morph_target;
 
 use crate::{
     primitives::Aabb,
@@ -11,6 +12,7 @@ use bevy_ecs::system::{lifetimeless::SRes, SystemParamItem};
 use bevy_math::*;
 use bevy_reflect::TypeUuid;
 use bevy_utils::EnumVariantMeta;
+pub use morph_target::*;
 use std::{borrow::Cow, collections::BTreeMap};
 use wgpu::{
     util::BufferInitDescriptor, BufferUsages, IndexFormat, PrimitiveTopology, VertexFormat,
@@ -18,91 +20,6 @@ use wgpu::{
 
 pub const INDEX_BUFFER_ASSET_INDEX: u64 = 0;
 pub const VERTEX_ATTRIBUTE_BUFFER_ID: u64 = 10;
-
-/// A [morph target] for a parent mesh. A given [`Mesh`] may have zero or more
-/// morph targets that affect the final rendered result.
-///
-/// [morph target]: https://en.wikipedia.org/wiki/Morph_target_animation
-#[derive(Debug, Clone)]
-pub struct MorphTarget {
-    pub position_displacement: Option<Vec<[f32; 3]>>,
-    pub normal_displacement: Option<Vec<[f32; 3]>>,
-    pub tangent_displacement: Option<Vec<[f32; 3]>>,
-}
-
-impl MorphTarget {
-    const VERTEX_SIZE: usize = std::mem::size_of::<f32>() * 8;
-
-    /// Counts all vertices of the morph target.
-    ///
-    /// # Panics
-    /// Panics if the attributes have different vertex counts.
-    pub fn count_vertices(&self) -> usize {
-        let mut vertex_count = self.position_displacement.as_ref().map(|v| v.len());
-        if let Some(ref attribute_data) = self.normal_displacement {
-            let attribute_len = attribute_data.len();
-            if let Some(previous_vertex_count) = vertex_count {
-                assert_eq!(previous_vertex_count, attribute_len,
-                        "Attribute {} has a different vertex count ({}) than other attributes ({}) in this morph target.", Mesh::ATTRIBUTE_NORMAL, attribute_len, previous_vertex_count);
-            }
-            vertex_count = Some(attribute_len);
-        }
-        if let Some(ref attribute_data) = self.tangent_displacement {
-            let attribute_len = attribute_data.len();
-            if let Some(previous_vertex_count) = vertex_count {
-                assert_eq!(previous_vertex_count, attribute_len,
-                        "Attribute {} has a different vertex count ({}) than other attributes ({}) in this morph target.", Mesh::ATTRIBUTE_TANGENT, attribute_len, previous_vertex_count);
-            }
-            vertex_count = Some(attribute_len);
-        }
-
-        vertex_count.unwrap_or(0)
-    }
-
-    pub fn get_vertex_buffer_data(&self) -> Vec<u8> {
-        let vertex_count = self.count_vertices();
-        let mut attributes_interleaved_buffer = vec![0; vertex_count * Self::VERTEX_SIZE];
-        // bundle into interleaved buffers
-        Self::interleave_bytes(
-            &self.position_displacement,
-            &mut attributes_interleaved_buffer,
-            3,
-            0,
-        );
-        Self::interleave_bytes(
-            &self.normal_displacement,
-            &mut attributes_interleaved_buffer,
-            2,
-            3,
-        );
-        Self::interleave_bytes(
-            &self.tangent_displacement,
-            &mut attributes_interleaved_buffer,
-            3,
-            5,
-        );
-        attributes_interleaved_buffer
-    }
-
-    fn interleave_bytes<T: bevy_core::Pod>(
-        src: &Option<Vec<T>>,
-        dst: &mut [u8],
-        float_count: usize,
-        float_offset: usize,
-    ) {
-        let attribute_offset = float_offset * std::mem::size_of::<f32>();
-        let attribute_size = float_count * std::mem::size_of::<f32>();
-        if let Some(attribute_values) = src.as_ref() {
-            let attributes_bytes = cast_slice(&attribute_values);
-            for (vertex_index, attribute_bytes) in
-                attributes_bytes.chunks_exact(attribute_size).enumerate()
-            {
-                let offset = vertex_index * Self::VERTEX_SIZE + attribute_offset;
-                dst[offset..offset + attribute_size].copy_from_slice(attribute_bytes);
-            }
-        }
-    }
-}
 
 // TODO: allow values to be unloaded after been submitting to the GPU to conserve memory
 #[derive(Debug, TypeUuid, Clone)]
@@ -113,7 +30,7 @@ pub struct Mesh {
     /// for this mesh. Attribute name maps to attribute values.
     /// Uses a [`BTreeMap`] because, unlike [`HashMap`], it has a defined iteration order,
     /// which allows easy stable vertex buffers (i.e. same buffer order)
-    /// 
+    ///
     /// [HashMap]: std::collections::HashMap
     attributes: BTreeMap<Cow<'static, str>, VertexAttributeValues>,
     morph_targets: Vec<MorphTarget>,
